requirements.txt
# Python Dependencies for Real-Time Data Pipeline

# Core Dependencies
python-dotenv==1.0.0

# Kafka
confluent-kafka==2.3.0
kafka-python==2.0.2

# ClickHouse
clickhouse-connect==0.6.23
clickhouse-driver==0.2.6

# MongoDB
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Machine Learning - Scikit-learn
scikit-learn==1.3.2
joblib==1.3.2

# Machine Learning - Spark
pyspark==3.5.0

# Machine Learning - Dask
dask[complete]==2023.12.0
dask-ml==2023.3.24

# Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Monitoring & Logging
loguru==0.7.2

# Utilities
requests==2.31.0
tqdm==4.66.1
schedule==1.2.0

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
ipython==8.18.1
jupyter==1.0.0

# Performance
psutil==5.9.6
